Add ElementArrangement in eager mode type conversion (#3601)
* Add ElementArrangement in eager mode

With unit tests.


* Refactor move EA_MAP to DtypeOpTable


* Refactor user builder function for EA_MAP


* Add rescale stl for dtype to runtime path

Refactor: move compile time path helper function into pass_utils so it
can be reused in runtime path _monkey_patch.py.


* Fix spyre->cpu EA & unblock D2D path


* Fix unit tests to use device combinations

Refactor: use table based tests


* Refactor use modes in pytest table

To avoid code repetition and reuse existing code for compile, eager,
fp16 & bf16 parameters.


* Increase ElementArrangement tests coverage

* Extend existing tests to cover both compile & eager modes.
* Cover all fp16 dtypes ["float16", "bfloat16"] in tests.
* Fix custom op `stagger_to_standard_ea` to include fp16 & bf16 types.
* Use clean readable test IDs.


---------

+++ torch_spyre/_inductor/customops.py
@@ -984,7 +984,7 @@ def stagger_to_standard_ea(x: torch.Tensor) -> torch.Tensor:
     # Each fp16 stick (64 elements) staggers independently with half=32.
     FP16_STICK = 64
     half = FP16_STICK // 2  # 32 — fixed, independent of n
-    P = torch.zeros(n, n, dtype=torch.float16, device="cpu")
+    P = torch.zeros(n, n, dtype=x.dtype, device="cpu")
     for phys_j in range(n):
         stick_base = (phys_j // FP16_STICK) * FP16_STICK
         local_phys = phys_j % FP16_STICK
